Files: simplify inheritance of file implementations
Still unresolved:
 - How can we specify the desired resolution order (and still allow it
   be flexible / extensible)?
