+++ ultratrace2/model/files/bundle.py
@@ -1,34 +1,41 @@
 import logging
 import os
 
-from typing import Dict, Sequence, FrozenSet
+from typing import Dict, FrozenSet, Optional, Sequence, Type
 
-from .impls import Sound, Alignment, ImageSet
+from .adt import AlignmentFile, ImageSetFile, SoundFile, TypedFile
 
 
 logger = logging.getLogger(__name__)
 
 
 class FileBundle:
-    def __init__(self, name: str):
+    def __init__(
+        self,
+        name: str,
+        alignment_file: Optional[AlignmentFile] = None,
+        image_file: Optional[ImageSetFile] = None,
+        sound_file: Optional[SoundFile] = None,
+    ):
         self.name = name
-        self.alignment_file = Alignment()
-        self.image_file = ImageSet()
-        self.sound_file = Sound()
-
-    def interpret(self, path: str):  # FIXME: add signature
-        return (
-            self.alignment_file.interpret(path)
-            or self.image_file.interpret(path)
-            or self.sound_file.interpret(path)
-        )  # noqa: E126
+        self.alignment_file = alignment_file
+        self.image_file = image_file
+        self.sound_file = sound_file
 
     def has_impl(self) -> bool:
-        return (
-            self.alignment_file.has_impl()
-            or self.image_file.has_impl()
-            or self.sound_file.has_impl()
+        return any(
+            f is not None
+            for f in [self.alignment_file, self.image_file, self.sound_file]
         )
+
+    def set_alignment_file(self, alignment_file: AlignmentFile) -> None:
+        self.alignment_file = alignment_file
+
+    def set_image_file(self, image_file: ImageSetFile) -> None:
+        self.image_file = image_file
+
+    def set_sound_file(self, sound_file: SoundFile) -> None:
+        self.sound_file = sound_file
 
     def __repr__(self):
         return f'Bundle("{self.name}",{self.alignment_file},{self.image_file},{self.sound_file})'
@@ -56,9 +63,9 @@ class FileBundleList:
         self.has_sound_impl: bool = False
 
         for bundle in bundles.values():
-            self.has_alignment_impl |= bundle.alignment_file.has_impl()
-            self.has_image_impl |= bundle.image_file.has_impl()
-            self.has_sound_impl |= bundle.sound_file.has_impl()
+            self.has_alignment_impl |= bundle.alignment_file is not None
+            self.has_image_impl |= bundle.image_file is not None
+            self.has_sound_impl |= bundle.sound_file is not None
 
     @classmethod
     def build_from_dir(
@@ -87,10 +94,24 @@ class FileBundleList:
                     )
                     continue
 
-                if name not in bundles:
-                    bundles[name] = FileBundle(name)
-
-                if not bundles[name].interpret(filepath):
+                impl_cls = guess_file_type(filepath)
+                if impl_cls is None:
                     logger.warning(f"unrecognized filetype: {filepath}")
+                    continue
+
+                if name not in bundles:
+                    bundles[name] = FileBundle(filepath)
+
+                file_impl = impl_cls(filepath)
+                if isinstance(file_impl, AlignmentFile):
+                    bundles[name].set_alignment_file(file_impl)
+                elif isinstance(file_impl, ImageSetFile):
+                    bundles[name].set_image_file(file_impl)
+                elif isinstance(file_impl, SoundFile):
+                    bundles[name].set_sound_file(file_impl)
 
         return cls(bundles)
+
+
+def guess_file_type(path: str) -> Optional[Type[TypedFile]]:
+    raise NotImplementedError()
