+++ ultratrace2/model/files/impls/__init__.py
@@ -1,3 +1,12 @@
-from .alignment import Alignment  # noqa: F401
-from .image import ImageSet  # noqa: F401
-from .sound import Sound  # noqa: F401
+# alignment files
+from .measurement import Measurement  # noqa: F401
+from .textgrid import TextGrid  # noqa: F401
+
+# imageset files
+from .dicom import DICOM  # noqa: F401
+
+# sound files
+from .flac import FLAC  # noqa: F401
+from .mp3 import MP3  # noqa: F401
+from .ogg import Ogg  # noqa: F401
+from .wav import WAV  # noqa: F401
