+++ ultratrace2/model/files/impls/flac.py
@@ -1,0 +1,8 @@
+from typing import ClassVar, Sequence
+
+from ..adt import TypedFile
+
+
+class FLAC(TypedFile):
+    mimetypes: ClassVar[Sequence[str]] = []
+    extensions: ClassVar[Sequence[str]] = []
