+++ ultratrace2/model/files/impls/mp3.py
@@ -1,0 +1,8 @@
+from typing import ClassVar, Sequence
+
+from ..adt import TypedFile
+
+
+class MP3(TypedFile):
+    mimetypes: ClassVar[Sequence[str]] = []
+    extensions: ClassVar[Sequence[str]] = []
