+++ ultratrace2/model/files/impls/ogg.py
@@ -1,0 +1,8 @@
+from typing import ClassVar, Sequence
+
+from ..adt import TypedFile
+
+
+class Ogg(TypedFile):
+    mimetypes: ClassVar[Sequence[str]] = []
+    extensions: ClassVar[Sequence[str]] = []
