+++ ultratrace2/model/files/impls/textgrid.py
@@ -1,0 +1,7 @@
+from ..adt import AlignmentFile
+
+
+class TextGrid(AlignmentFile):
+
+    mimetypes = ["text/plain"]
+    extensions = [".textgrid"]
