+++ ultratrace2/model/files/impls/wav.py
@@ -1,0 +1,6 @@
+from ..adt import TypedFile
+
+
+class WAV(TypedFile):
+    mimetypes = ["audio/x-wav", "audio/wav"]
+    extensions = ["wav"]
